code cleanup

+++ Hanekawa/Modules/Account/Account.cs
@@ -50,10 +50,10 @@ namespace Hanekawa.Modules.Account
                 var globalUserRank = db.AccountGlobals.CountAsync(x => x.TotalExp >= glUserData.TotalExp);
 
                 await Task.WhenAll(rank, total, globalRank, globalUserRank);
-                var nxtLevel = _calculate.GetServerLevelRequirement((int)userdata.Level);
+                var nxtLevel = _calculate.GetServerLevelRequirement(userdata.Level);
                 await Context.ReplyAsync(new EmbedBuilder()
                     .CreateDefault(Context.Guild.Id)
-                    .WithAuthor(new EmbedAuthorBuilder { Name = user.GetName() })
+                    .WithAuthor(new EmbedAuthorBuilder {Name = user.GetName()})
                     .WithThumbnailUrl(user.GetAvatar())
                     .WithFields(new List<EmbedFieldBuilder>
                     {
@@ -94,7 +94,8 @@ namespace Hanekawa.Modules.Account
                     rank++;
                 }
 
-                await PagedReplyAsync(pages.PaginateBuilder(Context.Guild.Id, Context.Guild, $"Level leaderboard for {Context.Guild.Name}",
+                await PagedReplyAsync(pages.PaginateBuilder(Context.Guild.Id, Context.Guild,
+                    $"Level leaderboard for {Context.Guild.Name}",
                     10));
             }
         }
@@ -156,6 +157,7 @@ namespace Hanekawa.Modules.Account
                 await Context.Channel.SendFileAsync(stream, "profile.png");
             }
         }
+
         /*
          <-- MAY USE THIS FOR LATER -->
          TODO: PREVIEW, MAYBE USE
